Update new tests to cypress
