+++ projects/ng-table-virtual-scroll/src/lib/table-item-size.directive.cy.ts
@@ -1,10 +1,40 @@
+import { DataSource } from '@angular/cdk/collections';
 import { CdkVirtualScrollViewport, ScrollingModule } from '@angular/cdk/scrolling';
-import { Component, ViewChild, ViewEncapsulation } from '@angular/core';
+import { CdkTableModule } from '@angular/cdk/table';
+import { Component, Type, ViewChild, ViewEncapsulation } from '@angular/core';
 import { MatTableModule } from '@angular/material/table';
 import { mount } from 'cypress/angular';
-import { TableVirtualScrollDataSource } from './table-data-source';
+import { CdkTableVirtualScrollDataSource, TableVirtualScrollDataSource } from './table-data-source';
 import { TableItemSizeDirective } from './table-item-size.directive';
 
+interface Data {
+  id: number;
+}
+
+abstract class TestComponent {
+  headerEnabled = true;
+  footerEnabled = false;
+
+  @ViewChild(CdkVirtualScrollViewport, { static: true })
+  viewport: CdkVirtualScrollViewport;
+
+  @ViewChild(TableItemSizeDirective, { static: true })
+  directive: TableItemSizeDirective;
+
+  displayedColumns = ['id'];
+
+  data = Array(50).fill(0).map((_, i) => ({ id: i }));
+  data2 = Array(50).fill(0).map((_, i) => ({ id: i + 50 }));
+
+  dataSource = new this.dataSourceClass(this.data);
+
+  protected constructor(protected dataSourceClass: Type<DataSource<Data>>) {
+  }
+
+  changeDataSource() {
+    this.dataSource = new this.dataSourceClass(this.data2);
+  }
+}
 
 @Component({
   template: `
@@ -68,34 +98,101 @@ import { TableItemSizeDirective } from './table-item-size.directive';
   `],
   encapsulation: ViewEncapsulation.None
 })
-class TableVirtualScrollComponent {
-  headerEnabled = true;
-  footerEnabled = false;
-
-  @ViewChild(CdkVirtualScrollViewport, { static: true })
-  viewport: CdkVirtualScrollViewport;
-
-  @ViewChild(TableItemSizeDirective, { static: true })
-  directive: TableItemSizeDirective;
-
-  displayedColumns = ['id'];
-
-  dataSource = new TableVirtualScrollDataSource(Array(50).fill(0).map((_, i) => ({ id: i })));
+class MatTableTestComponent extends TestComponent {
+  constructor() {
+    super(TableVirtualScrollDataSource);
+  }
+}
 
 
-  changeDataSource() {
-    this.dataSource = new TableVirtualScrollDataSource(Array(50).fill(0).map((_, i) => ({ id: i + 50 })));
+@Component({
+  template: `
+    <cdk-virtual-scroll-viewport tvsItemSize="10"
+                                 headerHeight="20"
+                                 footerHeight="15"
+                                 bufferMultiplier="0.5"
+                                 [headerEnabled]="headerEnabled"
+                                 [footerEnabled]="footerEnabled"
+                                 class="wrapper">
+
+      <table cdk-table [dataSource]="dataSource">
+
+        <ng-container *ngIf="headerEnabled">
+          <tr cdk-header-row *cdkHeaderRowDef="displayedColumns"></tr>
+        </ng-container>
+        <tr cdk-row *cdkRowDef="let row; columns: displayedColumns;"></tr>
+        <ng-container *ngIf="footerEnabled">
+          <tr cdk-footer-row *cdkFooterRowDef="displayedColumns"></tr>
+        </ng-container>
+
+        <ng-container cdkColumnDef="id">
+          <th cdk-header-cell *cdkHeaderCellDef>No.</th>
+          <td cdk-cell *cdkCellDef="let element">el - {{element.id}}</td>
+          <td cdk-footer-cell *cdkFooterCellDef class="footer-cell">End.</td>
+        </ng-container>
+
+      </table>
+
+    </cdk-virtual-scroll-viewport>
+    <button (click)="changeDataSource()" data-cy="changeDataSource"></button>
+  `,
+  styles: [`
+    .wrapper {
+      height: 40px;
+    }
+
+    tr {
+      height: auto !important;
+    }
+
+    th {
+      height: 20px !important;
+    }
+
+    td {
+      height: 10px !important;
+    }
+
+    th, td {
+      padding: 0 !important;
+      margin: 0 !important;
+      border-width: 0 !important;
+      border-style: none !important;
+      font-size: 8px;
+    }
+
+    .footer-cell {
+      height: 15px !important;
+    }
+  `],
+  encapsulation: ViewEncapsulation.None
+})
+class CdkTableTestComponent extends TestComponent {
+  constructor() {
+    super(CdkTableVirtualScrollDataSource);
   }
 }
 
 describe('TableItemSizeDirective', () => {
+  describe('CdkTable', () => {
+    runTableTests(CdkTableModule, CdkTableTestComponent);
+  });
+  describe('MatTable', () => {
+    runTableTests(MatTableModule, MatTableTestComponent);
+  });
+});
+
+function runTableTests(
+  tableModule: typeof CdkTableModule,
+  tableComponent: Type<TestComponent>
+) {
   describe('common actions', () => {
-    let testComponent: TableVirtualScrollComponent;
+    let testComponent: TestComponent;
     let viewport: CdkVirtualScrollViewport;
 
     beforeEach(() => {
-      mount(TableVirtualScrollComponent, {
-        imports: [ScrollingModule, MatTableModule],
+      mount(tableComponent, {
+        imports: [ScrollingModule, tableModule],
         declarations: [TableItemSizeDirective]
       })
         .then(mountResponse => {
@@ -143,8 +240,8 @@ describe('TableItemSizeDirective', () => {
 
   describe('initialization variants', () => {
     it('should have correct height with footer', () => {
-      mount(TableVirtualScrollComponent, {
-        imports: [ScrollingModule, MatTableModule],
+      mount(tableComponent, {
+        imports: [ScrollingModule, tableModule],
         declarations: [TableItemSizeDirective],
         componentProperties: {
           footerEnabled: true
@@ -159,8 +256,8 @@ describe('TableItemSizeDirective', () => {
     });
 
     it('should have correct height without header', () => {
-      mount(TableVirtualScrollComponent, {
-        imports: [ScrollingModule, MatTableModule],
+      mount(tableComponent, {
+        imports: [ScrollingModule, tableModule],
         declarations: [TableItemSizeDirective],
         componentProperties: {
           headerEnabled: false,
@@ -175,4 +272,4 @@ describe('TableItemSizeDirective', () => {
         });
     });
   });
-});
+}
